refactor: reorganize style directory structure
