+++ src/styles/components/Footer.scss
