+++ src/styles/components/NextBtn.scss
