+++ src/styles/components/Typing.scss
