+++ src/styles/pages/About.scss
