+++ src/styles/pages/Education.scss
