+++ src/styles/pages/Project.scss
