+++ src/styles/pages/Skill.scss
